api/src/payments/pays/views.py
from rest_framework import generics

from pays.models import Payment
from pays.serializers import PaymentSerializer


class PaymentsView(generics.ListCreateAPIView):
    queryset = Payment.objects.all()
    serializer_class = PaymentSerializer


class PaymentView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Payment.objects.all()
    serializer_class = PaymentSerializer
